+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Versions
 
+- [0.3.0](#030)
 - [0.2.0](#020)
 - [0.1.0](#010)
 
@@ -13,10 +14,19 @@
 
 <br>
 
+### 0.3.0
+
+- Modular dependencies updated
+- Cleanup
+- Updated types
+- Refactoring
+
+<br>
+
 ### 0.2.0
 
 - More config options
-- better plugin integration
+- Better plugin integration
 - Updated types
 
 <br>
@@ -24,5 +34,5 @@
 ### 0.1.0
 
 - Lib setup
-- basic engine
+- Basic engine
 - Testing system